Delete cache only if it exists.
git-svn-id: http://svn.browsershots.org/branches/shotfactory-django@1595 ca065b8f-b50a-0410-be8c-a9c8883ec1d6

+++ shotfactory04/gui/linux/konqueror.py
@@ -45,5 +45,6 @@ class Gui(base.Gui):
         for cachedir in glob(os.path.join(dotdir, 'cache-*')):
             while os.path.islink(cachedir):
                 cachedir = os.readlink(cachedir)
-            print 'deleting cache', cachedir
-            shutil.rmtree(cachedir)
+            if os.path.exists(cachedir):
+                print 'deleting cache', cachedir
+                shutil.rmtree(cachedir)
